Add files via upload
The copy.ps1 file will copy all files located in the "Documents" folder of the currently logged in user, and save them to a USB drive.

The main issue at the moment is the external drive is called "D:\", if a user has any extra drives the USB will not show up as D:\. This means the files will not be saved to desired USB. I need a more generic way of addressing the USB drive. This will allow it to be used more seamlessly.